refactor: split named call argument lowering
